+++ docs/experiments/raven-target/CoreDeclarations.cs
@@ -33,6 +33,7 @@ static class CoreDeclarations
             using var image = Mono.Cecil.AssemblyDefinition.ReadAssembly(stream);
             var module = image.MainModule;
             CalendarBindings.ProjectLayout(module);
+            if (libraryBootstrap) PrimitiveLibrary.Project(module);
             NamespaceFunctions.ProjectMath(module);
             NamespaceFunctions.ProjectFault(module);
             var unit = module.GetType("System.PropagationUnit");